requirements.txt
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
docopt==0.6.2
idna==2.7
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.5
pyparsing==2.2.0
python-dateutil==2.7.3
python-ternary==1.0.4
pytz==2018.4
requests==2.19.1
six==1.11.0
tqdm==4.23.4
urllib3==1.23
yarg==0.1.9
